+++ ja2lib/Strategic/StrategicMines.h
@@ -164,7 +164,7 @@ UINT8 GetHeadMinerIndexForMine(INT8 bMineIndex);
 UINT16 GetHeadMinerProfileIdForMine(INT8 bMineIndex);
 
 // Find the sector location of a mine
-void GetMineSector(UINT8 ubMineIndex, INT16* psX, INT16* psY);
+void GetMineSector(UINT8 ubMineIndex, u8* psX, u8* psY);
 
 void IssueHeadMinerQuote(INT8 bMineIndex, UINT8 ubQuoteType);
